+++ firmware/boron-ins-fsm/src/stateMachine.h
@@ -10,17 +10,22 @@
 
 // ascii table goes up to 7F, so pick something greater than that
 // which is also unlikely to be part of a door ID or a threshold/timer const
-#define INITIALIZE_STATE_MACHINE_CONSTS_FLAG           0x8888
-#define INITIALIZE_STATE3_MAX_LONG_STILLNESS_TIME_FLAG 0x8888
-#define INITIALIZE_STATE0_OCCUPANT_DETECTION_FLAG      0x8888
+// try to make them different, first 3 have already been written into memory on some live sensors
+#define INITIALIZE_STATE_MACHINE_CONSTS_FLAG                0x8888
+#define INITIALIZE_STATE4_HIGH_CONF_MAX_STILLNESS_TIME_FLAG 0x8888
+#define INITIALIZE_STATE0_OCCUPANT_DETECTION_FLAG           0x8888
+#define INITIALIZE_HIGH_CONF_INS_THRESHOLD_FLAG             0x9999
 
 // initial (default) values for state machine, can be changed via console function
 // or by writing something other than 0x8888 to the above flag in flash
-#define INS_THRESHOLD                   60
-#define STATE0_OCCUPANT_DETECTION_TIMER 172800000  // 2 days
-#define STATE1_MAX_TIME                 15000      // ms = 15s
-#define STATE2_MAX_DURATION             1200000    // ms = 20 min
-#define STATE3_MAX_STILLNESS_TIME       120000     // ms = 2 minutes
+#define LOW_CONF_INS_THRESHOLD              80
+#define HIGH_CONF_INS_THRESHOLD             1          // disabled by default
+#define STATE0_OCCUPANT_DETECTION_MAX_TIME  120000     // 2 min
+#define STATE1_MAX_TIME                     5000       // 5s
+#define STATE2_MAX_DURATION                 1200000    // 20 min
+#define STATE3_LOW_CONF_MAX_STILLNESS_TIME  120000     // 2 minutes
+#define STATE4_HIGH_CONF_MAX_STILLNESS_TIME 60000      // 1 min
+#define LOW_CONF_STILLNESS_TIME             3000000    // 5 min
 
 // How often to publish Heartbeat messages
 #define SM_HEARTBEAT_INTERVAL 660000  // ms = 11 min
@@ -49,6 +54,7 @@ void state0_idle();
 void state1_15sCountdown();
 void state2_duration();
 void state3_stillness();
+void state4_true_stillness();
 
 void publishDebugMessage(int, unsigned char, float, unsigned long);
 void publishStateTransition(int, int, unsigned char, float);
@@ -69,17 +75,20 @@ extern StateHandler stateHandler;
 extern unsigned long state1_timer;
 extern unsigned long state2_duration_timer;
 extern unsigned long state3_stillness_timer;
+extern unsigned long state4_true_stillness_timer;
 
 // state machine constants stored in flash
-extern unsigned long ins_threshold;
-extern unsigned long state0_occupant_detection_timer;
+extern unsigned long low_conf_ins_threshold;
+extern unsigned long high_conf_ins_threshold;
+extern unsigned long state0_occupant_detection_max_time;
 extern unsigned long state1_max_time;
 extern unsigned long state2_max_duration;
-extern unsigned long state3_max_stillness_time;
-extern unsigned long state3_max_long_stillness_time;
+extern unsigned long state3_low_conf_max_stillness_time;
+extern unsigned long state4_high_conf_max_stillness_time;
 
 // whether or not the current session has sent alerts
 extern bool hasDurationAlertBeenSent;
 extern bool hasStillnessAlertBeenSent;
+extern bool hasTrueStillnessAlertBeenSent;
 
 #endif